Removed dependency on dgc in favor of the new Slash Commands offered by the Discord API.
closes #1

+++ kardbot/utils.go
@@ -8,7 +8,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-const MaxDiscordMsgLen uint = 2000
+const MaxDiscordMsgLen uint64 = 2000
 
 // Some characters are optional when matching the bot name.
 // This function returns a regexp string to appropriately
@@ -44,9 +44,9 @@ func buildRegexAltGroup(alts []string) string {
 
 // Return a non-negative random number in the inclusive range [min, max].
 // If max <= min, returns the maximum uint value and an error.
-func randFromRange(min, max uint) (uint, error) {
+func randFromRange(min, max uint64) (uint64, error) {
 	if max <= min {
-		return ^uint(0), fmt.Errorf("max (%d) cannot be less than or equal to min (%d)", max, min)
+		return ^uint64(0), fmt.Errorf("max (%d) cannot be less than or equal to min (%d)", max, min)
 	}
-	return uint(rand.Intn(int((max-min)+1))) + min, nil
+	return uint64(rand.Intn(int((max-min)+1))) + min, nil
 }
